requirements.txt
advertools==0.13.5
aiohttp==3.9.2
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
build==1.0.3
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.23
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
commonmark==0.9.1
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.2
cssselect==1.2.0
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.28.1
googleapis-common-protos==1.62.0
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.21.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
incremental==22.10.0
ipydatawidgets==4.3.5
ipython==8.18.1
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonpickle==3.0.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kubernetes==29.0.0
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.9.0
openlimit==0.3.0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parsel==1.8.1
parso==0.8.3
peewee==3.17.0
pexpect==4.9.0
pillow==10.2.0
plotly==5.18.0
posthog==3.4.2
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==4.25.3
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyDispatcher==2.0.7
Pygments==2.17.2
pyOpenSSL==24.0.0
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
pyvis==0.3.2
PyYAML==6.0.1
queuelib==1.6.2
regex==2023.12.25
replicate==0.24.0
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
rsa==4.9
Scrapy==2.11.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
tldextract==5.1.1
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.1
traitlets==5.14.2
traittypes==0.2.1
Twisted==22.10.0
twython==3.9.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.8
urllib3==1.26.18
uvicorn==0.27.1
uvloop==0.19.0
w3lib==2.1.2
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
